Fixed something about prices, used a float to parse it instead of an int like the first time and then multiplied by 100

+++ mysite/main/views.py
@@ -90,7 +90,8 @@ def GamesSearchHandler(response, searchBy: str, games : list):
         if priceStr is None or priceStr == "":
             return
         
-        price = int(float(priceStr) * 100)
+
+        price = float(priceStr) * 100
 
 
         Games_res = select("Games",columns="id, Name, Short_description, Base_price, Current_price, Coming_soon, Release_Date",
